RUM-4098: Move datastore interfaces to api
