+++ dd-sdk-android-core/api/apiSurface
@@ -102,7 +102,7 @@ interface com.datadog.android.api.feature.FeatureContextUpdateReceiver
 interface com.datadog.android.api.feature.FeatureEventReceiver
   fun onReceive(Any)
 interface com.datadog.android.api.feature.FeatureScope
-  var dataStore: com.datadog.android.core.persistence.datastore.DataStoreHandler
+  var dataStore: com.datadog.android.api.storage.datastore.DataStoreHandler
   fun withWriteContext(Boolean = false, (com.datadog.android.api.context.DatadogContext, com.datadog.android.api.storage.EventBatchWriter) -> Unit)
   fun sendEvent(Any)
   fun <T: Feature> unwrap(): T
@@ -152,6 +152,15 @@ data class com.datadog.android.api.storage.RawBatchEvent
   constructor(ByteArray, ByteArray = EMPTY_BYTE_ARRAY)
   override fun equals(Any?): Boolean
   override fun hashCode(): Int
+interface com.datadog.android.api.storage.datastore.DataStoreCallback<T: Any>
+  fun onSuccess(com.datadog.android.core.persistence.datastore.DataStoreContent<T>?)
+  fun onFailure()
+interface com.datadog.android.api.storage.datastore.DataStoreHandler
+  fun <T: Any> setValue(String, T, Int = 0, com.datadog.android.core.persistence.Serializer<T>)
+  fun <T: Any> value(String, Int = 0, DataStoreCallback<T>, com.datadog.android.core.internal.persistence.Deserializer<String, T>)
+  fun removeValue(String)
+  companion object 
+    const val CURRENT_DATASTORE_VERSION: Int
 interface com.datadog.android.core.InternalSdkCore : com.datadog.android.api.feature.FeatureSdkCore
   val networkInfo: com.datadog.android.api.context.NetworkInfo
   val trackingConsent: com.datadog.android.privacy.TrackingConsent
@@ -291,18 +300,8 @@ interface com.datadog.android.core.persistence.Serializer<T: Any>
   fun serialize(T): String?
   companion object 
 fun <T: Any> Serializer<T>.serializeToByteArray(T, com.datadog.android.api.InternalLogger): ByteArray?
-interface com.datadog.android.core.persistence.datastore.DataStoreCallback<T: Any>
-  fun onSuccess(DataStoreContent<T>)
-  fun onFailure()
-  fun onNoData()
 data class com.datadog.android.core.persistence.datastore.DataStoreContent<T: Any>
   constructor(Int, T?)
-interface com.datadog.android.core.persistence.datastore.DataStoreHandler
-  fun <T: Any> setValue(String, T, Int = 0, com.datadog.android.core.persistence.Serializer<T>)
-  fun <T: Any> value(String, Int = 0, DataStoreCallback<T>, com.datadog.android.core.internal.persistence.Deserializer<String, T>)
-  fun removeValue(String)
-  companion object 
-    const val CURRENT_DATASTORE_VERSION: Int
 class com.datadog.android.core.sampling.RateBasedSampler : Sampler
   constructor(() -> Float)
   constructor(Float)
