+++ dd-sdk-android-core/api/dd-sdk-android-core.api
@@ -325,9 +325,9 @@ public abstract interface class com/datadog/android/api/feature/FeatureEventRece
 }
 
 public abstract interface class com/datadog/android/api/feature/FeatureScope {
-	public abstract fun getDataStore ()Lcom/datadog/android/core/persistence/datastore/DataStoreHandler;
+	public abstract fun getDataStore ()Lcom/datadog/android/api/storage/datastore/DataStoreHandler;
 	public abstract fun sendEvent (Ljava/lang/Object;)V
-	public abstract fun setDataStore (Lcom/datadog/android/core/persistence/datastore/DataStoreHandler;)V
+	public abstract fun setDataStore (Lcom/datadog/android/api/storage/datastore/DataStoreHandler;)V
 	public abstract fun unwrap ()Lcom/datadog/android/api/feature/Feature;
 	public abstract fun withWriteContext (ZLkotlin/jvm/functions/Function2;)V
 }
@@ -462,6 +462,28 @@ public final class com/datadog/android/api/storage/RawBatchEvent {
 	public final fun getMetadata ()[B
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
+}
+
+public abstract interface class com/datadog/android/api/storage/datastore/DataStoreCallback {
+	public abstract fun onFailure ()V
+	public abstract fun onSuccess (Lcom/datadog/android/core/persistence/datastore/DataStoreContent;)V
+}
+
+public abstract interface class com/datadog/android/api/storage/datastore/DataStoreHandler {
+	public static final field CURRENT_DATASTORE_VERSION I
+	public static final field Companion Lcom/datadog/android/api/storage/datastore/DataStoreHandler$Companion;
+	public abstract fun removeValue (Ljava/lang/String;)V
+	public abstract fun setValue (Ljava/lang/String;Ljava/lang/Object;ILcom/datadog/android/core/persistence/Serializer;)V
+	public abstract fun value (Ljava/lang/String;ILcom/datadog/android/api/storage/datastore/DataStoreCallback;Lcom/datadog/android/core/internal/persistence/Deserializer;)V
+}
+
+public final class com/datadog/android/api/storage/datastore/DataStoreHandler$Companion {
+	public static final field CURRENT_DATASTORE_VERSION I
+}
+
+public final class com/datadog/android/api/storage/datastore/DataStoreHandler$DefaultImpls {
+	public static synthetic fun setValue$default (Lcom/datadog/android/api/storage/datastore/DataStoreHandler;Ljava/lang/String;Ljava/lang/Object;ILcom/datadog/android/core/persistence/Serializer;ILjava/lang/Object;)V
+	public static synthetic fun value$default (Lcom/datadog/android/api/storage/datastore/DataStoreHandler;Ljava/lang/String;ILcom/datadog/android/api/storage/datastore/DataStoreCallback;Lcom/datadog/android/core/internal/persistence/Deserializer;ILjava/lang/Object;)V
 }
 
 public abstract interface class com/datadog/android/core/InternalSdkCore : com/datadog/android/api/feature/FeatureSdkCore {
@@ -791,12 +813,6 @@ public final class com/datadog/android/core/persistence/SerializerKt {
 	public static final fun serializeToByteArray (Lcom/datadog/android/core/persistence/Serializer;Ljava/lang/Object;Lcom/datadog/android/api/InternalLogger;)[B
 }
 
-public abstract interface class com/datadog/android/core/persistence/datastore/DataStoreCallback {
-	public abstract fun onFailure ()V
-	public abstract fun onNoData ()V
-	public abstract fun onSuccess (Lcom/datadog/android/core/persistence/datastore/DataStoreContent;)V
-}
-
 public final class com/datadog/android/core/persistence/datastore/DataStoreContent {
 	public fun <init> (ILjava/lang/Object;)V
 	public final fun component1 ()I
@@ -808,23 +824,6 @@ public final class com/datadog/android/core/persistence/datastore/DataStoreConte
 	public final fun getVersionCode ()I
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
-}
-
-public abstract interface class com/datadog/android/core/persistence/datastore/DataStoreHandler {
-	public static final field CURRENT_DATASTORE_VERSION I
-	public static final field Companion Lcom/datadog/android/core/persistence/datastore/DataStoreHandler$Companion;
-	public abstract fun removeValue (Ljava/lang/String;)V
-	public abstract fun setValue (Ljava/lang/String;Ljava/lang/Object;ILcom/datadog/android/core/persistence/Serializer;)V
-	public abstract fun value (Ljava/lang/String;ILcom/datadog/android/core/persistence/datastore/DataStoreCallback;Lcom/datadog/android/core/internal/persistence/Deserializer;)V
-}
-
-public final class com/datadog/android/core/persistence/datastore/DataStoreHandler$Companion {
-	public static final field CURRENT_DATASTORE_VERSION I
-}
-
-public final class com/datadog/android/core/persistence/datastore/DataStoreHandler$DefaultImpls {
-	public static synthetic fun setValue$default (Lcom/datadog/android/core/persistence/datastore/DataStoreHandler;Ljava/lang/String;Ljava/lang/Object;ILcom/datadog/android/core/persistence/Serializer;ILjava/lang/Object;)V
-	public static synthetic fun value$default (Lcom/datadog/android/core/persistence/datastore/DataStoreHandler;Ljava/lang/String;ILcom/datadog/android/core/persistence/datastore/DataStoreCallback;Lcom/datadog/android/core/internal/persistence/Deserializer;ILjava/lang/Object;)V
 }
 
 public final class com/datadog/android/core/sampling/RateBasedSampler : com/datadog/android/core/sampling/Sampler {
